requirements.txt
streamlit
pillow
requests
